refactor(delivery): establish delivery/ — channel_key, router, readiness
Slice 5, ROI-ranked at minimal blast radius (owner directive): the delivery/
package is the doctrine's how/when/where seam. result_channel_key (1 importer),
result_router (3), result_ready (5). All three are stdlib-only leaves. The
ag2-sparrow vendor map now points at delivery/readiness.py (vendored name
unchanged); the readiness delegation contract accepts the dotted src form and
the vendored .result_ready form.
