+++ packages/ag2-sparrow/tools/sync_from_src.py
@@ -28,7 +28,7 @@ MAP = {
     "task_archive.py": "task_archive.py",
     "local_task_protocol.py": "local_task_protocol.py",
     "result_markers.py": "result_markers.py",
-    "result_ready.py": "result_ready.py",
+    "delivery/readiness.py": "result_ready.py",
     "dedup_recovery.py": "dedup_recovery.py",
     "workspace_lock.py": "workspace_lock.py",
     "chat_secret_filter.py": "chat_secret_filter.py",
