+++ src/delivery/readiness.py
