+++ tests/result-ready-contract.test.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-"""Contract for src/result_ready.py and delegation by every delivery consumer.
+"""Contract for src/delivery/readiness.py and delegation by every delivery consumer.
 
 Readiness of a task-result file has one owner. Each consumer binds its own
 results directory and keeps only provider-specific delivery.
@@ -15,7 +15,7 @@ from pathlib import Path
 REPO = Path(__file__).resolve().parent.parent
 sys.path.insert(0, str(REPO / "src"))
 
-from result_ready import is_ready_body, read_ready_result  # noqa: E402
+from delivery.readiness import is_ready_body, read_ready_result  # noqa: E402
 
 # Every consumer that decides "is this result ready to deliver?".
 CONSUMERS = {
@@ -92,7 +92,7 @@ class DelegationTest(unittest.TestCase):
                 self.assertTrue(path.exists(), f"{name}: missing at {path}")
                 self.assertRegex(
                     path.read_text(),
-                    r"from (?:\.)?result_ready import read_ready_result",
+                    r"from (?:delivery\.readiness|\.result_ready) import read_ready_result",
                     f"{name}: does not import read_ready_result from the shared owner",
                 )
 
@@ -107,14 +107,14 @@ class DelegationTest(unittest.TestCase):
                 self.assertEqual(
                     hits, [],
                     f"{name}: reads a result file directly ({hits}) — readiness "
-                    f"belongs to src/result_ready.read_ready_result",
+                    f"belongs to src/delivery/readiness.read_ready_result",
                 )
 
     def test_sparrow_bundle_matches_src(self):
         pkg = (REPO / "packages" / "ag2-sparrow" / "ag2_sparrow" / "result_ready.py")
         self.assertTrue(pkg.exists(), "result_ready.py not bundled into ag2-sparrow")
         self.assertEqual(
-            pkg.read_text(), (REPO / "src" / "result_ready.py").read_text(),
+            pkg.read_text(), (REPO / "src" / "delivery" / "readiness.py").read_text(),
             "ag2-sparrow copy drifted from src/ — run tools/sync_from_src.py",
         )
